Improve handling of the fenced code blocks
Supports whitespace after the fence. ex:

``` plantuml
stuff
```

Support more than 3 ticks or tildes for code blocks, this is still
only a part of the variations in the wilds [1]

Finds the correct matching ending fence by using identation,
this required adding some more groups to the regexp and
the option MULTILINE. ex.

    ```` plantuml
        stuff containing ```
    ````

[1] - https://pandoc.org/MANUAL.html#fenced-code-blocks

Closes: #34