Change how fence blocks are processed
Use a fixed regex to match all fence blocks, not just the
ones that are modified by this plugin.

de-indent the code block before processing.

This will avoid having wrong matches within a fenced block. ex:

````
Hey, this is just regular text, presenting a plantuml code block.

``` plantuml
@startuml
...
@enduml
```
````

De-indentind the code is necessary if the code block is in a structural
element like a list.
(Apart from being not correct, handling from_file_prefix did not work in
that case)